+++ src/main/java/com/sparta/taptoon/domain/portfolio/dto/portfolioDto/response/GetPortfolioResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.taptoon.domain.portfolio.dto.response;
+package com.sparta.taptoon.domain.portfolio.dto.portfolioDto.response;
 
 import com.sparta.taptoon.domain.member.entity.Member;
 import com.sparta.taptoon.domain.portfolio.entity.Portfolio;
@@ -7,10 +7,10 @@ import lombok.Builder;
 import java.time.LocalDateTime;
 
 @Builder
-public record CreatePortfolioResponse(Member member, String title, String content, String fileUrl, LocalDateTime createdAt) {
+public record GetPortfolioResponse(Member member, String title, String content, String fileUrl, LocalDateTime createdAt) {
 
-    public static CreatePortfolioResponse from(Portfolio portfolio) {
-        return CreatePortfolioResponse.builder()
+    public static GetPortfolioResponse from(Portfolio portfolio) {
+        return GetPortfolioResponse.builder()
                 .member(portfolio.getMember())
                 .title(portfolio.getTitle())
                 .content(portfolio.getContent())
