+++ src/main/java/com/sparta/taptoon/domain/portfolio/service/PortfolioService.java
@@ -2,11 +2,11 @@ package com.sparta.taptoon.domain.portfolio.service;
 
 import com.sparta.taptoon.domain.member.entity.Member;
 import com.sparta.taptoon.domain.member.repository.MemberRepository;
-import com.sparta.taptoon.domain.portfolio.dto.request.PortfolioRequest;
-import com.sparta.taptoon.domain.portfolio.dto.response.CreatePortfolioResponse;
-import com.sparta.taptoon.domain.portfolio.dto.response.GetAllPortfolioResponse;
-import com.sparta.taptoon.domain.portfolio.dto.response.GetPortfolioResponse;
-import com.sparta.taptoon.domain.portfolio.dto.response.UpdatePortfolioResponse;
+import com.sparta.taptoon.domain.portfolio.dto.portfolioDto.request.PortfolioRequest;
+import com.sparta.taptoon.domain.portfolio.dto.portfolioDto.response.CreatePortfolioResponse;
+import com.sparta.taptoon.domain.portfolio.dto.portfolioDto.response.GetAllPortfolioResponse;
+import com.sparta.taptoon.domain.portfolio.dto.portfolioDto.response.GetPortfolioResponse;
+import com.sparta.taptoon.domain.portfolio.dto.portfolioDto.response.UpdatePortfolioResponse;
 import com.sparta.taptoon.domain.portfolio.entity.Portfolio;
 import com.sparta.taptoon.domain.portfolio.repository.PortfolioRepository;
 import jakarta.transaction.Transactional;
@@ -90,5 +90,27 @@ public class PortfolioService {
 
         // 수정한 포트폴리오 반환
         return UpdatePortfolioResponse.from(portfolio);
+    }
+
+    // 포트폴리오 삭제
+    @Transactional
+    public void removePortfolio(Long portfolioId, Long memberId) {
+
+        // 유저 조회
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new RuntimeException("유저 정보를 찾을 수 없습니다."));
+
+        // 수정할 포트폴리오 Id로 찾기
+        Portfolio portfolio = portfolioRepository.findById(portfolioId)
+                .orElseThrow(() -> new RuntimeException("포트폴리오를 찾을 수 없습니다."));
+
+        // 수정할 포트폴리오가 유저의 포트폴리오인지 검사
+        if(!portfolio.getMember().equals(member)) {
+            throw new IllegalArgumentException("다른 사람의 포트폴리오를 수정할 수 없습니다.");
+        }
+
+        // 포트폴리오 삭제 (boolean type true로 변경)
+        portfolio.remove();
+
     }
 }
